requirements.txt
asgiref==3.2.7
attrs==19.3.0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
django>=2.1.0
djangorestframework>=3.8.0
docutils==0.16
idna==2.9
jeepney==0.4.3
jsonschema>=3.0.0
keyring==21.2.1
packaging==20.3
pkginfo==1.5.0.1
pycparser==2.20
pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
pyyaml==5.3.1
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
secretstorage==3.1.2
six==1.14.0
sqlparse==0.3.1
tqdm==4.46.0
twine==3.1.1
urllib3==1.25.9
webencodings==0.5.1
wheel==0.34.2
